@@ -18,12 +18,12 @@ package com.example.android.mygarden.ui;
 
 import android.content.Context;
 import android.database.Cursor;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.android.mygarden.R;
 import com.example.android.mygarden.provider.PlantContract.PlantEntry;
